Separate the extension namespaces for extending Cpg and Step class.
This became necessary because we otherwise ended up with the Step
extension methods as method on the Cpg class. Since this does not make
sense it did not compile but was suggest in code completion and more
important it ended up in the docs.
