+++ query-primitives/src/main/scala/io/shiftleft/codepropertygraph/ext/queryprimitives/package.scala
@@ -1,10 +1,10 @@
-package io.shiftleft.queryprimitives.steps.ext
+package io.shiftleft.codepropertygraph.ext
 
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.queryprimitives.steps.starters.NodeTypeStarters
 
 package object queryprimitives {
-
   implicit def toNodeTypeStarters(cpg: Cpg): NodeTypeStarters =
     new NodeTypeStarters(cpg)
+
 }
