+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/steps/Steps.scala
@@ -22,7 +22,6 @@ import shapeless.ops.product.ToHList
   */
 class Steps[NodeType, Labels <: HList](val raw: GremlinScala.Aux[NodeType, Labels])
     extends ext.Enrichable
-    with ext.queryprimitives.Enrichable
     with ext.securityprofile.Enrichable
     with ext.semanticcpg.Enrichable
     with ext.dataflowengine.Enrichable {
