+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/steps/ext/securityprofile/Enrichable.scala
